# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Bancaria de Transferencias",
    "description": "API REST para operaciones bancarias de transferencias, diseñada para fines académicos de Ingeniería de Servicios Web.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Servidor local de desarrollo" }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/cuentas": {
      "get": {
        "tags": ["Cuentas"],
        "summary": "Lista las cuentas del cliente autenticado",
        "description": "Retorna un arreglo con las cuentas asociadas al cliente autenticado. Se pueden filtrar por estado y moneda.",
        "operationId": "listarCuentasCliente",
        "parameters": [
          { "name": "estado", "in": "query", "description": "Estado de la cuenta (ej. ACTIVA, INACTIVA)", "required": false, "schema": { "type": "string", "example": "ACTIVA" } },
          { "name": "moneda", "in": "query", "description": "Moneda de la cuenta", "required": false, "schema": { "type": "string", "example": "DOP", "enum": ["DOP", "USD", "EUR"] } }
        ],
        "responses": {
          "200": {
            "description": "Lista de cuentas del cliente",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Cuenta" } }
              }
            }
          },
          "401": {
            "description": "No autorizado o token inválido",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/cuentas/{numeroCuenta}": {
      "get": {
        "tags": ["Cuentas"],
        "summary": "Consulta detalle y saldo de una cuenta específica",
        "description": "Retorna el detalle completo y saldo de la cuenta indicada por número.",
        "operationId": "obtenerDetalleCuenta",
        "parameters": [
          { "name": "numeroCuenta", "in": "path", "required": true, "schema": { "type": "string", "minLength": 10, "maxLength": 25, "example": "1001-2345-6789-0001" } }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la cuenta",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cuenta" } } }
          },
          "404": {
            "description": "Cuenta no encontrada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "No autorizado o token inválido",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/beneficiarios": {
      "get": {
        "tags": ["Beneficiarios"],
        "summary": "Lista beneficiarios asociados al cliente",
        "description": "Retorna un arreglo con los beneficiarios registrados por el cliente autenticado.",
        "operationId": "listarBeneficiarios",
        "responses": {
          "200": {
            "description": "Lista de beneficiarios",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Beneficiario" } } } }
          },
          "401": {
            "description": "No autorizado o token inválido",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "post": {
        "tags": ["Beneficiarios"],
        "summary": "Registra un nuevo beneficiario",
        "description": "Permite al cliente registrar un nuevo beneficiario para transferencias.",
        "operationId": "crearBeneficiario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BeneficiarioRequest" },
              "example": {
                "alias": "Madre",
                "nombreBeneficiario": "María Pérez",
                "numeroCuenta": "1001-8888-7777-0002",
                "banco": "Banco Demo",
                "documentoIdentidad": "00198765432"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Beneficiario creado exitosamente",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Beneficiario" } } }
          },
          "400": {
            "description": "Error de validación en datos de entrada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } }
          },
          "409": {
            "description": "Beneficiario ya existe",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/beneficiarios/{idBeneficiario}": {
      "put": {
        "tags": ["Beneficiarios"],
        "summary": "Actualiza alias o datos permitidos del beneficiario",
        "description": "Permite modificar el alias o datos permitidos de un beneficiario existente.",
        "operationId": "actualizarBeneficiario",
        "parameters": [
          { "name": "idBeneficiario", "in": "path", "required": true, "schema": { "type": "string", "example": "BEN-001" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/beneficiarios_idBeneficiario_body" },
              "example": { "alias": "Madre Actualizada" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Beneficiario actualizado correctamente",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Beneficiario" } } }
          },
          "400": {
            "description": "Error de validación en datos de entrada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } }
          },
          "404": {
            "description": "Beneficiario no encontrado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "delete": {
        "tags": ["Beneficiarios"],
        "summary": "Elimina un beneficiario",
        "description": "Permite eliminar un beneficiario registrado.",
        "operationId": "eliminarBeneficiario",
        "parameters": [
          { "name": "idBeneficiario", "in": "path", "required": true, "schema": { "type": "string", "example": "BEN-001" } }
        ],
        "responses": {
          "204": { "description": "Beneficiario eliminado correctamente" },
          "404": {
            "description": "Beneficiario no encontrado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/api/admin/transferencias/sospechosas": {
      "get": {
        "tags": ["Administración"],
        "summary": "Lista operaciones marcadas como sospechosas",
        "description": "Permite a usuarios con rol ADMIN_BANCO o AUDITOR consultar transferencias marcadas para revisión.",
        "operationId": "listarTransferenciasSospechosas",
        "parameters": [
          { "name": "fechaDesde", "in": "query", "required": false, "schema": { "type": "string", "format": "date", "example": "2026-01-01" } },
          { "name": "fechaHasta", "in": "query", "required": false, "schema": { "type": "string", "format": "date", "example": "2026-02-17" } },
          { "name": "estadoRevision", "in": "query", "required": false, "schema": { "type": "string", "example": "EN_REVISION" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de transferencias sospechosas",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TransferenciaDetalle" } } } }
          },
          "401": { "description": "No autorizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Acceso denegado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/transferencias/{idTransaccion}/revision": {
      "patch": {
        "tags": ["Administración"],
        "summary": "Aprueba o rechaza una transferencia sospechosa",
        "description": "Permite a un usuario con rol ADMIN_BANCO aprobar o rechazar una transferencia marcada para revisión.",
        "operationId": "revisarTransferenciaSospechosa",
        "parameters": [
          { "name": "idTransaccion", "in": "path", "required": true, "schema": { "type": "string", "example": "TRX-20260216-0002" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RevisionTransferenciaRequest" },
              "example": { "accion": "APROBAR", "comentario": "Validación manual completada" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transferencia revisada correctamente",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransferenciaDetalle" } } }
          },
          "401": { "description": "No autorizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Acceso denegado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Transferencia no encontrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "codigo": { "type": "string", "example": "ERROR_INTERNO" },
          "mensaje": { "type": "string", "example": "Ha ocurrido un error" },
          "detalles": { "type": "object", "nullable": true },
          "timestamp": { "type": "string", "format": "date-time" },
          "path": { "type": "string" },
          "requestId": { "type": "string" }
        }
      },
      "ValidationErrorResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ErrorResponse" },
          {
            "type": "object",
            "properties": {
              "codigo": { "type": "string", "example": "ERROR_VALIDACION" },
              "detalles": { "type": "object", "example": { "monto": ["El monto debe ser mayor que 0"] } }
            }
          }
        ]
      },
...
      "Cuenta": {
        "type": "object",
        "properties": {
          "numeroCuenta": { "type": "string", "minLength": 10, "maxLength": 25, "example": "1001-2345-6789-0001" },
          "tipoCuenta": { "type": "string", "enum": ["AHORRO", "CORRIENTE"], "example": "AHORRO" },
          "moneda": { "type": "string", "enum": ["DOP", "USD", "EUR"], "example": "DOP" },
          "saldoDisponible": { "type": "number", "format": "double", "example": 50250.75 },
          "saldoContable": { "type": "number", "format": "double", "example": 51000.75 },
          "estado": { "type": "string", "enum": ["ACTIVA", "INACTIVA"], "example": "ACTIVA" },
          "titular": { "$ref": "#/components/schemas/Cuenta_titular" }
        }
      },
      "Cuenta_titular": {
        "type": "object",
        "properties": {
          "idCliente": { "type": "string", "example": "CLI-1001" },
          "nombreCompleto": { "type": "string", "example": "Juan Pérez" }
        }
      },
      "Beneficiario": {
        "type": "object",
        "properties": {
          "idBeneficiario": { "type": "string", "example": "BEN-001" },
          "alias": { "type": "string", "example": "Madre" },
          "nombreBeneficiario": { "type": "string", "example": "María Pérez" },
          "numeroCuenta": { "type": "string", "minLength": 10, "maxLength": 25, "example": "1001-8888-7777-0002" },
          "banco": { "type": "string", "example": "Banco Demo" },
          "documentoIdentidad": { "type": "string", "example": "00198765432" }
        }
      },
      "BeneficiarioRequest": {
        "required": ["alias", "nombreBeneficiario", "numeroCuenta", "banco", "documentoIdentidad"],
        "type": "object",
        "properties": {
          "alias": { "type": "string", "example": "Madre" },
          "nombreBeneficiario": { "type": "string", "example": "María Pérez" },
          "numeroCuenta": { "type": "string", "minLength": 10, "maxLength": 25, "example": "1001-8888-7777-0002" },
          "banco": { "type": "string", "example": "Banco Demo" },
          "documentoIdentidad": { "type": "string", "example": "00198765432" }
        }
      },
      "beneficiarios_idBeneficiario_body": {
        "type": "object",
        "properties": {
          "alias": { "type": "string", "example": "Madre Actualizada" },
          "nombreBeneficiario": { "type": "string", "example": "María Pérez Actualizada" },
          "numeroCuenta": { "type": "string", "minLength": 10, "maxLength": 25 },
          "banco": { "type": "string" },
          "documentoIdentidad": { "type": "string" }
        }
      },
...
      "TransferenciaResponse": {
        "type": "object",
        "properties": {
          "idTransaccion": { "type": "string", "example": "TRX-20260217-0001" },
          "estado": { "type": "string", "enum": ["COMPLETADA", "EN_REVISION", "RECHAZADA"], "example": "COMPLETADA" },
          "fecha": { "type": "string", "format": "date-time" },
          "monto": { "type": "number", "format": "double", "example": 5000 },
          "moneda": { "type": "string", "enum": ["DOP", "USD", "EUR"], "example": "DOP" },
          "concepto": { "type": "string", "example": "Pago factura" },
          "cuentaOrigen": { "type": "string", "example": "1001-2345-6789-0001" },
          "cuentaDestino": { "type": "string", "example": "1001-9876-5432-1000" }
        }
      },
      "TransferenciaDetalle": {
        "allOf": [
          { "$ref": "#/components/schemas/TransferenciaResponse" },
          {
            "type": "object",
            "properties": {
              "idBeneficiario": { "type": "string", "nullable": true, "example": "BEN-001" },
              "mensaje": { "type": "string", "example": "Transferencia procesada correctamente" }
            }
          }
        ]
      },
...
      "RevisionTransferenciaRequest": {
        "required": ["accion"],
        "type": "object",
        "properties": {
          "accion": { "type": "string", "enum": ["APROBAR", "RECHAZAR"], "example": "APROBAR" },
          "comentario": { "type": "string", "nullable": true, "example": "Validación manual completada" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token"
      }
    }
  }
}